+++ resources.go
@@ -1366,9 +1366,26 @@ func Provider() tfbridge.ProviderInfo {
 			"aws_ram_principal_association": {Tok: awsResource(ramMod, "PrincipalAssociation")},
 			"aws_ram_resource_association":  {Tok: awsResource(ramMod, "ResourceAssociation")},
 			// Relational Database Service (RDS)
-			"aws_rds_cluster":          {Tok: awsResource(rdsMod, "Cluster")},
+			"aws_rds_cluster": {
+				Tok: awsResource(rdsMod, "Cluster"),
+				Fields: map[string]*tfbridge.SchemaInfo{
+					"engine": {
+						Type: awsResource(rdsMod, "EngineType"),
+					},
+					"engine_mode": {
+						Type: awsResource(rdsMod, "EngineMode"),
+					},
+				},
+			},
 			"aws_rds_cluster_endpoint": {Tok: awsResource(rdsMod, "ClusterEndpoint")},
-			"aws_rds_cluster_instance": {Tok: awsResource(rdsMod, "ClusterInstance")},
+			"aws_rds_cluster_instance": {
+				Tok: awsResource(rdsMod, "ClusterInstance"),
+				Fields: map[string]*tfbridge.SchemaInfo{
+					"engine": {
+						Type: awsResource(rdsMod, "EngineType"),
+					},
+				},
+			},
 			"aws_rds_cluster_parameter_group": {
 				Tok: awsResource(rdsMod, "ClusterParameterGroup"),
 				Fields: map[string]*tfbridge.SchemaInfo{
@@ -1962,6 +1979,12 @@ func Provider() tfbridge.ProviderInfo {
 						DestFiles: []string{
 							"runtimes.ts", // a union type and constants for available Lambda runtimes.
 							"lambdaMixins.ts",
+						},
+					},
+					"rds": {
+						DestFiles: []string{
+							"engineMode.ts",
+							"engineType.ts",
 						},
 					},
 					"s3": {